refactor(components): begin removing unnecessary core compoonents
