+++ packages/core/components/defaultLoading/spinner/index.js
