--- packages/core/components/helpers/container/container.css
@@ -1,40 +1,0 @@
-.container {
-  margin-left: auto;
-  margin-right: auto;
-  width: 100%;
-}
-
-.padding {
-  padding-left: 1.25rem;
-  padding-right: 1.25rem;
-}
-
-.xxl {
-  composes: container;
-  composes: padding;
-  max-width: 90rem;
-}
-
-.xl {
-  composes: container;
-  composes: padding;
-  max-width: 80rem;
-}
-
-.lg {
-  composes: container;
-  composes: padding;
-  max-width: 64rem;
-}
-
-.md {
-  composes: container;
-  composes: padding;
-  max-width: 48rem;
-}
-
-.sm {
-  composes: container;
-  composes: padding;
-  max-width: 32rem;
-}
