--- packages/core/components/helpers/container/readme.md
@@ -1,9 +1,0 @@
-Container helper component. This component is useful for constraining the width of its children, centering them, and adding outer padding.
-
-```js
-<Container size="sm">
-    <div style={{ backgroundColor: 'orange' }}>
-        this is a container
-    </div>
-</Container>
-```
