--- packages/core/components/helpers/heading/heading.css
@@ -1,61 +1,0 @@
-.heading {
-  font-weight: 700;
-}
-
-.heading a {
-  color: inherit;
-  text-decoration: none;
-}
-
-.heading em {
-  font-style: italic;
-}
-
-.heading strong {
-  font-weight: 700;
-}
-
-.step-down-two {
-  composes: heading;
-  font-size: 10px;
-}
-
-.step-down-one {
-  composes: heading;
-  font-size: 14px;
-}
-
-.base {
-  composes: heading;
-  font-size: 16px;
-}
-
-.step-up-one {
-  composes: heading;
-  font-size: 20px;
-}
-
-.step-up-two {
-  composes: heading;
-  font-size: 24px;
-}
-
-.step-up-three {
-  composes: heading;
-  font-size: 34px;
-}
-
-.step-up-four {
-  composes: heading;
-  font-size: 48px;
-}
-
-.step-up-five {
-  composes: heading;
-  font-size: 60px;
-}
-
-.step-up-six {
-  composes: heading;
-  font-size: 96px;
-}
