--- packages/core/components/helpers/helpers.md
@@ -1,3 +1,0 @@
-Irving offers a collection of components for helping you with common tasks. Generally speaking, these helper components are intended to correspond to particular styles, making it easier to standardize and manage your site's design:
-* Heading, Button, and Container components for helping standardize various versions/layouts of each of those things.
-* A Link component for distingushing absolute and relative links and handling them accordingly. Relative links will use history push, absolute links will not.
