--- packages/core/components/image/image.css
@@ -1,44 +1,0 @@
-.wrapper {
-  display: block;
-}
-
-.img,
-.picture {
-  display: block;
-}
-
-.lazyload .placeholder {
-  display: block;
-}
-
-.lazyload .img,
-.lazyload .picture {
-  display: none;
-}
-
-.lazyload.loaded .placeholder {
-  display: none;
-}
-
-.lazyload.loaded .img,
-.lazyload.loaded .picture {
-  display: block;
-}
-
-.apsectRatio {
-  height: 0;
-  position: relative;
-}
-
-.apsectRatio picture,
-.apsectRatio img {
-  height: auto;
-  left: 0;
-  position: absolute;
-  top: 0;
-  width: 100%;
-}
-
-.apsectRatio .placeholder {
-  height: 100%;
-}
