--- packages/core/components/placeholder/placeholder.css
@@ -1,28 +1,0 @@
-.wrapper {
-  margin-bottom: 2rem;
-  width: 100%;
-}
-
-.wrapper pre {
-  font-family: monospace;
-}
-
-.heading2 {
-  font-size: 32px;
-}
-
-.heading3 {
-  font-size: 28px;
-}
-
-.heading4 {
-  font-size: 24px;
-}
-
-.heading5 {
-  font-size: 20px;
-}
-
-.children {
-  margin-left: 32px;
-}
