--- packages/core/components/rawHTML/rawHTML.css
@@ -1,12 +1,0 @@
-.wrapper {
-
-  i,
-  em {
-    font-style: italic;
-  }
-
-  b,
-  strong {
-    font-weight: 700;
-  }
-}
