--- packages/core/components/socialItem/readme.md
@@ -1,9 +1,0 @@
-Component for displaying a single icon/link corresponding, generally, to a social media service. It could, however, be used for any icon that should link out somewhere.
-
-```js
-<SocialItem
-    type="facebook"
-    url="facebook.com/irving"
-    displayIcon={true}
-/>
-```
