--- packages/core/components/socialItem/socialItem.css
@@ -1,16 +1,0 @@
-.wrapper {
-  display: block;
-}
-
-.wrapper svg {
-  height: 22px;
-  width: 22px;
-}
-
-.link {
-  align-items: center;
-  display: flex;
-  height: 40px;
-  justify-content: center;
-  width: 40px;
-}
