--- packages/core/components/socialList/socialList.css
@@ -1,8 +1,0 @@
-.wrapper {
-  display: block;
-}
-
-.list {
-  align-items: center;
-  display: flex;
-}
